Support for C99 intermingling of declaration and statements, patch by Benedikt Freisen.
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@10450 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/SDCCast.h
@@ -50,7 +50,7 @@ typedef struct ast
   unsigned lvalue:1;
   unsigned initMode:1;
   unsigned reversed:1;
-  int level;                    /* level for expr */
+  long level;                   /* level for expr */
   int block;                    /* block number   */
   int seqPoint;                 /* sequence point */
   /* union of values expression can have */
